Another instrumentation skipping for clone syscall
Similarly to what was done in commit 26fc73e57a6d53, we needed to escape
the clone (56) syscall to prevent crashes, and we use a previous
instruction blacklist for that.

This works in practice but it is not respecting assembly semantics.
For robustness, we MUST at some point remove these quirks and start to
only instrument the syscalls we really want to wrap.