Re-base libsystrap onto librunt, and tidy up
This eliminates elfutils.[ch], whose role is now handled by librunt. We
also tidy up the use of raw syscalls to minimise the amount of libc
header-hackery and header-avoidance required, splitting off sigill.c and
raw-syscalls-impl.h. And we use musl as our supporting self-contained
libc, rather than a glibc build (which takes ages, and feels more
fragile to interface with).

Currently there is a segfault bug somewhere in the ld.so during libxed's
lazy linking. I am digging into this but need to keep hacksawing in
parallel, so committing these changes now.