Makefile cleanup for multiarch
We now build in a build dir that is named after our arch.

We are going to have to have multiarch versions of our dependencies;
here we add scaffolding in contrib/ but we do not complete the job, so
the build contines to be single-arch for now.

To make all this easier, we drop the non-xed decoder dependencies