do-syscall.h: fix clone alignment issues
We now use an out-of-line helper function to copy the stack, which also
returns to the calling asm code the new stack pointer to use for the
child thread. It also takes care of alignment issues, which were causing
"bad sigframe" kernel errors (delivering SIGSEGV), I think when the
XSAVE area of the fp state was not aligned properly (64 bytes). This
approach is so much cleaner than before. Some more cleanup needed, and
32-bit x86 is broken.