trap.c: explicitly detect vdso mapping and avoid instrumenting it
Somewhere along the line, the kernel has started allowing us to mprotect
the vdso and modify its text, in some cases. So we had both the 'fake'
and the 'real' vdso containing ud2s. That is wrong and created an
infinite loop of signal handling. So explicitly detect a vdso mapping
and decline to instrument it. Only the fake vdso should be instrumented.